Tests: Changes for POC to work
