+++ tests/unit/accordion/options.js
@@ -1,5 +1,5 @@
 define( [
-	"lib/qunit"
+	"lib/qunit",
 	"jquery",
 	"./helper",
 	"ui/accordion"
